re #1541: New security setup. Now there is a separate filter to connect to the apis authorization server and one that creates a SecurityContext based on the VerifyTokenResponse of the authorization server.
The unit tests had to be changed for the new situation. Unit tests that test the user in role action, are temporarily ignored, because at the moment no role based security is implemented.

Obsolete classes are deleted and now we use the resource server library of the apis project for the classes needed to verify the access tokens.