re #1552: mail support added. Administrators get e-mailed when a new user is logged on to their VRE. Users get and e-mail when their user-account is changed by the administrator. To configure the mail sending the config.xml now contains some extra configuration items.
To make it possible for a VRE to retrieve the user information of unverified users, a new role is added UNVERIFIED_USER. A new method is added to the UserResource. This method will only work with the security working. Because of this new method the RolesPartiallyAllowed annotation could be removed. Also the post-method on the UserResource is removed because users are created when they first log on to a VRE.

Some changes are made to search users by role. The MongoObjectMapper class now supports the use of Collections that contain wrapper classes or string.