Move v3 to own subdirectory
The poms were not changed, so this commit won't mvn compile, but the
alternative was a commit including both a few thousand moves _and_ some
changes. this way you can verify that this commit only moves things
around using

    git diff -M --name-status <this hash>^ <this hash> | grep -v ^R100

Sorry for breaking your git bisect flow. If the parent commit fails
then the move did not cause the break. If the parent commit succeeds
and the child commit (which fixes the build) does fail then the move is
the cause of your break.