Execute rmlMappings in dependency sorted order
We can execute a mapping by first importing the collection that is on
the one side of a one-to-many relation or by first importing the many
side.

First importing the one side allows for better error messages (I can
say "value foo is does not exist" at the spot of value foo instead of a
blurb with "values foo and bar were missing" at the end of the sheet.

This code will try to order the imports in such a way that the one side
is always imported first. If the mapping contains circulair
dependencies then it will only import those the other way around.