redirect from old (versionless) urls to v2.1
The old API is different from the new v2.1 API, which is why the url's
changed. However, we still receive requests on the old URL's from
handle id's and crawler bots. They give a 404 now which is worse than
a 200. No actual API clients exist that use the old API.