invert Vres and DataAccess dependency
The main difference can be seen in TimbuctooV4. At first the vres
instance was sent to the DataAccess class, and the Vres would contain
code to load itself from the database. Now the Vres is passed an
instance of DataAccess instead so I can move to Vres loading code over
to that class.