Remove finished migrations
They depend on GraphWrapper and make refactoring harder. There's a case
to be made for keeping them, but given the amount of code we're
changing we'd need to copy them and the files they depend on to a
separate package anyway. Therefore if we ever need to rerun them again
we might better check out an old version of timbuctoo.