prepare testGraphBuilder for with<dto> calls
Currently testGraphBuilder accepts withVertex and this means that each
unittest that creates a graph has to implement the vertices according
to the expectations of itself and all cooperating classes. For example:
if I want to change the behaviour of a getEntity method to use labels
some unittests will fail because they prepared a graph without labels,
even though all vertices have labels in the production db.

This prevents us from writing tests at the integration level, so we're
forced to skip to full launch-a-server integration tests too soon, or
we have to pay a high cost for changing code.

This changes allows me to implement GraphFragmentBuilders for say, a
VRE or an Entity so that I can change the way they are persisted to
the database at a central place.

This change consists of moving more logic to the VertexBuilder so that
the new GraphFragmentBuilder can be a simple interface on top of it.