Remove invalid spec
This spec ascertains if the current properties of the entity are
copied when a new collection is added, however you can have different
values for different collections so it is undefined behaviour what
values you en up with.

Multiple collections per vertex within the same VRE has never
consciously been implemented.

The proper fix is to remove namespacing from the properties (and use
different vertices for different VRE's)