handle upload async and return status updates
Uploads can take a while and the browser will then think that the
server has crashed and reset the connection. This new approach first
uploads all the data, then sends a small response to the browser (so
the connection moves from waiting, to recieved first byte) and then
processes the data. Sending status updates to the browser while doing
so, so that the frontend can present the user with a nice progress
dialog