Partially handle sameAs relations
Vertices now contain an extra property called rdfAlternatives.
Asserting a owl:sameAs relation results in a vertex with the
rdfAlternatives filled with those urls. (same-as in timbuctoo therefore
really means "the same as".

The case where both vertices already exist
and have different properties is explicitly not handled (by logging an
error). The case where the same property of same-as entities is
asserted to a different value after the same-as assertion has been
handled is silently not handled (last write wins)

I need this for importing many2many tables, because r2rml can only
handle these using dataset local uris. Whereas we also want them to
have uri's that do not contain the unique key in the spreadsheet. The
spreadsheet might not even have a unique key. Therefore we generate our
own key and if we do a many2many mapping we also generate the
spreadsheet specific keys using sameAs