remove run state
These files are created by timbuctoo while running. They should not be
in source control, but launching timbuctoo without them won't work so I
also made the run script autogenerate them.

I have also renamed the run script to make it clear that it is not
really for production runs (it launches in debugger mode by default,
and redefines the data dirs)