Implement cyclebreaker differently
The previous algorithm would only select the optimal split-off point if
you passed the input list in the right order. It now first detects
cycles, then selects the cycles that appear most often and breaks them.
We can change the selection procedure later on to take the preference
of a datasource into account (a large datasource might like it less to
be streamed twice then a small datasource)