make clean verify work if the source dir is dirty
At the end of the run, the concordion results are pasted back to the
source dir (the mounted dir containing the timbuctoo checkout from your
pc) if it does not have a target folder (i.e. you had a clean checkout)
it should create the target folder. If it does have a target folder
(i.e. you have run mvn package at least once) it should ignore the fact
that a target folder already exists.