Fix excel upload
In f2518f2936ce7868487849a9242d069ea2e87902 I moved the resultReporter
calls to the importer, so that the caller of Importer was no longer
responsible for calling ResultReporter at the right time.

I accidentally replaced a call to ResultReporter#startentity with one
to importer#startentity.

The side effects of this fix is that empty rows will no longer be
counted