Make FulltextIndexChangeListener use displaynames
Some artifact of history makes the fullTextIndex use a custom
configuration per collection. These are mostly the same as the
displayname except for wwdocuments where it also contains the
authorname and wwlocations where it erroneously uses a non-existing
property.

This change reworks the FulltextIndexChangeListener to generate the
indexes using the displayname with custom logic for wwdocuments. In the
next commit the classes that have become superfluous will be removed