stop typesHelper from using onUpdate
I agree that the AddLabelChangeListener should contain all the code
that adds/removes labels for types. The onUpdate method is pretty
specific for the crudservice/datastorehandler though and it's fit for
from the rdf importer is accidental. I therefore gave the RDF importer
it's own dedicated methods in the same class.