log duration correctly for streaming responses
The writeTo for a streaming response is called immediately. Even though
the streaming response has not closed the output yet. By wrapping the
stream we make sure to wait until the close of the http-response to log
the message.

This adds a new problem, namely that the MDC might have been reset by
another http requests when this request finished. So we send the MDC
context as a hashmap to the stream so the close log matches the open
log.

With this approach I can verify that logging works even if a subthread
is started by running an upload and simultaneously firing a lot of GET
requests at timbuctoo. As you can see both the httplog and a log
request from the subthread contain the correct request_id, while the
other request that is handled by thread dw-63 has its own request_id.

- INFO   >     PUT http://localhost:8080/v2.1/bulk-upload 70cfa0a0-d5e4-4ceb-a74e-311ee19f5143 dw-63 - PUT /v2.1/bulk-upload
[..SNIP...]
- INFO   >     GET http://localhost:8080/v2.1/system/vres f3219b1a-9ffd-489d-bd10-49804164f9e0 dw-63 - GET /v2.1/system/vres
[..SNIP...]
- INFO   < 200 GET http://localhost:8080/v2.1/system/vres (450 bytes) (43 ms) f3219b1a-9ffd-489d-bd10-49804164f9e0 dw-63 - GET /v2.1/system/vres
[..SNIP...]
- INFO   < 200 PUT http://localhost:8080/v2.1/bulk-upload (345 bytes) (4482 ms) 70cfa0a0-d5e4-4ceb-a74e-311ee19f5143 Thread-27
- INFO   deleted tempfile /var/folders/5m/1p490dy52tqcwb8j1d3qbzhw0000gn/T/file1204270350866471027.tmp after import 70cfa0a0-d5e4-4ceb-a74e-311ee19f5143 Thread-27