Log TTFB seperately from full duration
TTFB = Time To first byte. This is the moment that the "slow" spinner
in the browser turns into a "fast" spinner. It's best to have a quick
overall duration, but if that is not possible then it's better to have
a quick TTFB because from that moment on the frontend can start to show
feedback to the user (if written correctly)