group datafetchers and datafetcherstores together
There's now a clear entrypoint (The DataFetcherFactory implementation)
and that is the only point that is exposed outside of the package.

Well... TripleStore also implements EntityProvider which is how the
SchemaStore get's updated. I think we should make schemaStore keep
track of it's own data. But that would require writing twice as much to
the database which would currently make the import twice as slow. (and
duplicate all the data) so we'll leave that for now...