implement inverse relations correctly
Previous implementation was to append _inverse to the predicate and
store it like a normal relation, but then have some checks on whether a
predicate ended with _inverse.

This implementation stores predicates together with a token that
indicates the direction of the relation.

This commit also fixes a bug in SchemaEntityProcessor where an inverse
predicate that was sometimes used as a list and sometimes not would be
marked as a list depending on how it was used last (see
SchemaEntityProcessor line 77)