clean up DataSetFactory
It can now return a DataSet interface instead of only the fields of
that interface. I also move the "constructor function" in the factory
to the constructor in the interface. The factory is not really a
factory and will be renamed in a later commit.

You can now use a DataSet as a lock e.g.

....DataSet dataSet = dataSetFactory.getDataSet(userId, dataSetId);
    synchronized(dataSet) {
      try(Stream<CursorQuad> quads = dataSet.getQuadStore.getQuads()) {
        ...
      }
      dataSet.getImportManager().addLog()//a log created from the quads
    }