make json-ld identical to graphql schema
The graphql schema contains a {value} object because values cannot be
part of a union type. The generated json-ld simplified these objects for
known json types (string, integer, number etc). This change marks
graphql introspection values as a different categorie from rdf values so
that we can generate simple json values for introspected graphql types
and expanded objects for all rdf types. So a query for `{value}` will
return and object with a value key.